+++ order/src/main/java/com/klp/order/infrastructure/event/event/OrderPaidEvent.java
@@ -8,8 +8,8 @@ import java.util.UUID;
 public record OrderPaidEvent(
     UUID orderId,
     String idempotencyKey,
-    Long supplierId,
-    Long customerId,
+    UUID supplierId,
+    Long userId,
     List<DeliveryItem> items,
     LocalDateTime occurredAt
 ) {
@@ -37,7 +37,7 @@ public record OrderPaidEvent(
             order.getOrderId(),
             idempotencyKey,
             order.getSupplierId(),
-            order.getCustomerId(),
+            order.getUserId(),
             items,
             LocalDateTime.now()
         );
